Move stream classes to seperate files
